+++ src/Formatter.js
@@ -20,12 +20,15 @@ function parseFormatStr(formatStr) {
       opts.forEach((opt) => {
         if (!opt) return;
         const [key, ...rest] = opt.split(':');
-        const val = rest.join(':');
+        const val = rest
+          .join(':')
+          .trim()
+          .replace(/^'+|'+$/g, ''); // trim and replace ''
 
-        if (!formatOptions[key.trim()]) formatOptions[key.trim()] = val.trim();
-        if (val.trim() === 'false') formatOptions[key.trim()] = false;
-        if (val.trim() === 'true') formatOptions[key.trim()] = true;
-        if (!isNaN(val.trim())) formatOptions[key.trim()] = parseInt(val.trim(), 10);
+        if (!formatOptions[key.trim()]) formatOptions[key.trim()] = val;
+        if (val === 'false') formatOptions[key.trim()] = false;
+        if (val === 'true') formatOptions[key.trim()] = true;
+        if (!isNaN(val)) formatOptions[key.trim()] = parseInt(val, 10);
       });
     }
   }