+++ src/Interpolator.js
@@ -145,7 +145,7 @@ class Interpolator {
         const safeValue = todo.safeValue(value);
         str = str.replace(match[0], safeValue);
         if (skipOnVariables) {
-          todo.regex.lastIndex += safeValue.length;
+          todo.regex.lastIndex += value.length;
           todo.regex.lastIndex -= match[0].length;
         } else {
           todo.regex.lastIndex = 0;