+++ src/utils.js
@@ -136,7 +136,8 @@ export function escape(data) {
 
 export const isIE10 =
   typeof window !== 'undefined' &&
-  window.navigator &&
+  window.navigator && // https://github.com/i18next/i18next/issues/1778
+  typeof window.navigator.userAgentData === 'undefined' &&
   window.navigator.userAgent &&
   window.navigator.userAgent.indexOf('MSIE') > -1;