+++ test/backend/backendMock.js
@@ -12,9 +12,41 @@ class Backend {
   read(language, namespace, callback) {
     if (!this.retries[language]) this.retries[language] = 0;
 
-    if (namespace === 'retry' && this.retries[language] < 2) {
+    if (namespace.indexOf('fail') === 0) {
+      return callback('failed loading', true);
+    } else if (namespace === 'retry0') {
       this.retries[language]++;
       return callback('failed loading', true);
+    } else if (namespace === 'retry1' && this.retries[language] < 1) {
+      this.retries[language]++;
+      return callback('failed loading', true);
+    } else if (namespace === 'retry2' && this.retries[language] < 2) {
+      this.retries[language]++;
+      return callback('failed loading', true);
+    } else if (namespace === 'retry5' && this.retries[language] < 5) {
+      this.retries[language]++;
+      return callback('failed loading', true);
+    } else if (namespace === 'retry6' && this.retries[language] < 6) {
+      this.retries[language]++;
+      return callback('failed loading', true);
+    } else if (namespace === 'retry7' && this.retries[language] < 7) {
+      this.retries[language]++;
+      return callback('failed loading', true);
+
+      // // Is a retry, but not set to fail after a specific
+      // } else if (namespace.indexOf('retry') === 0) {
+
+      // }
+    } else if (namespace.indexOf('concurrentlyLonger') === 0) {
+      setTimeout(() => {
+        callback(null, { status: 'ok', namespace });
+      }, 400);
+      return;
+    } else if (namespace.indexOf('concurrently') === 0) {
+      setTimeout(() => {
+        callback(null, { status: 'ok', namespace });
+      }, 200);
+      return;
     } else {
       callback(null, { status: 'nok', retries: this.retries[language] });
       delete this.retries[language];
@@ -28,7 +60,7 @@ class Backend {
 
     if (!this.retries[language]) this.retries[language] = 0;
 
-    if (namespace === 'retry' && this.retries[language] < 2) {
+    if (namespace === 'retry2' && this.retries[language] < 2) {
       this.retries[language]++;
       return callback('failed loading', true);
     } else {
@@ -40,5 +72,7 @@ class Backend {
     }
   }
 }
+
+Backend.type = 'backend';
 
 export default Backend;