+++ test/i18next.spec.js
@@ -101,6 +101,12 @@ describe('i18next', () => {
         expect(t('key')).to.equal('ici!');
         expect(t.keyPrefix).to.equal('deeply.nested');
       });
+      it('should apply keyPrefix and also be able to overwrite it', () => {
+        i18next.addResource('it', 'translation', 'deeply.nested.key', 'quì!');
+        const t = i18next.getFixedT('it', null, 'deeply.nested');
+        expect(t('nested.key', { keyPrefix: 'deeply' })).to.equal('quì!');
+        expect(t.keyPrefix).to.equal('deeply.nested');
+      });
     });
   });