+++ test/typescript/init.test.ts
@@ -213,6 +213,9 @@ const data = i18next.getDataByLanguage('de');
 // verify the data.translation field exists
 console.log(data ? data.translation.myKey : 'data does not exist');
 
+// verify the field for a specific namespace exists
+console.log(data ? data.common.myKey : 'data does not exist');
+
 // or fix the namespace to anotherNamespace
 const anotherNamespace = i18next.getFixedT(null, 'anotherNamespace');
 const x: string = anotherNamespace('anotherNamespaceKey'); // no need to prefix ns i18n.t('anotherNamespace:anotherNamespaceKey');
@@ -461,16 +464,6 @@ i18next.init({
   },
   backend: locizeOptions,
   locizeLastUsed: locizeOptions,
-  editor: {
-    ...locizeOptions,
-    onEditorSaved: async (lng, ns) => {
-      // reload that namespace in given language
-      await i18next.reloadResources(lng, ns);
-      // trigger an event on i18n which triggers a rerender
-      // based on bindI18n below in react options
-      i18next.emit('editorSaved');
-    },
-  },
   react: {
     bindI18n: 'languageChanged editorSaved',
   },